Final version in prod

+++ PDBDownloader/Program.cs
@@ -223,15 +223,16 @@ namespace PDBDownloader
             ManageFile<ResultItem> bestFile = new ManageFile<ResultItem>(bestFileName);
             foreach (ResultItem item in bestFile.FindObjects())
             {
-                string outputDir = ConfigurationManager.AppSettings["outDirForDownload"] + item.filename + ".mmcif";
-                string apiUrl = "https://files.rcsb.org/download/" + item.filename + ".cif";
-                HttpResponseMessage response = await client.GetAsync(apiUrl);
-                response.EnsureSuccessStatusCode();
-                byte[] fileContent = await response.Content.ReadAsByteArrayAsync();
-                File.WriteAllBytes(outputDir, fileContent);
-                Console.WriteLine("Download with success : " + item.filename + ".mmcif");
-            }
-            
+                if(!File.Exists(ConfigurationManager.AppSettings["outDirForDownload"] + item.filename + ".mmcif"))
+                {
+                    string outputDir = ConfigurationManager.AppSettings["outDirForDownload"] + item.filename + ".mmcif";
+                    string apiUrl = "https://files.rcsb.org/download/" + item.filename + ".cif";
+                    HttpResponseMessage response = await client.GetAsync(apiUrl);
+                    byte[] fileContent = await response.Content.ReadAsByteArrayAsync();
+                    File.WriteAllBytes(outputDir, fileContent);
+                    Console.WriteLine("Download with success : " + item.filename + ".mmcif");
+                }               
+            }  
         }
 
         static void KeepBestData()
